chore: improve piece centric evaluation

+++ include/chesstillo/position.hpp
@@ -90,7 +90,7 @@ private:
   friend struct EvalState;
 
   friend int SEE(Position &position);
-  friend int Evaluate(Position &position);
+  friend float Evaluate(Position &position);
   friend std::vector<Move> GenerateMoves(Position &position);
   friend Bitboard CheckMask(Position &position);
   friend std::tuple<Bitboard, Bitboard> PinMask(Position &position);
